rename: auction-item 폴더 정리
