+++ apps/web/src/components/common/auction-item/card.tsx
